move component to file
